allocation: Rename task_allocation package to allocation
