+++ mrs/allocation/bidder.py
@@ -3,10 +3,10 @@ import logging
 
 from mrs.exceptions.task_allocation import NoSTPSolution
 from mrs.robot_base import RobotBase
-from mrs.structs.allocation import FinishRound
-from mrs.structs.allocation import TaskAnnouncement, Allocation
+from mrs.allocation.allocation import FinishRound
+from mrs.allocation.allocation import TaskAnnouncement, Allocation
 from mrs.structs.bid import Bid
-from mrs.task_allocation.bidding_rule import BiddingRule
+from mrs.allocation.bidding_rule import BiddingRule
 from fmlib.models.tasks import Task
 from ropod.structs.task import TaskStatus as TaskStatusConst
