+++ mrs/allocation/bidding_rule.py
