+++ mrs/config/mrta.py
@@ -1,6 +1,6 @@
-from mrs.task_allocation.auctioneer import Auctioneer
+from mrs.allocation.auctioneer import Auctioneer
 from mrs.task_execution.dispatcher import Dispatcher
-from mrs.task_allocation.bidder import Bidder
+from mrs.allocation.bidder import Bidder
 from mrs.task_execution.schedule_monitor import ScheduleMonitor
 import logging
 from stn.stp import STP
