+++ setup.py
@@ -5,7 +5,7 @@ from setuptools import setup
 setup(name='mrs',
       packages=['mrs', 'mrs.config', 'mrs.config.builders', 'mrs.config.default',
                 'mrs.db.models', 'mrs.db.models.performance', 'mrs.exceptions',
-                'mrs.experiments', 'mrs.structs', 'mrs.task_allocation',
+                'mrs.experiments', 'mrs.structs', 'mrs.allocation',
                 'mrs.task_execution', 'mrs.utils'],
       version='0.2.0',
       description='Multi-Robot System (MRS) components for performing'
